refactored all components to be pure
As we are using pure components and plan to use them with
immutable and persistent data structures we can make the
react components pure. This allows react to have a more efficient
render strategy. There can never be changes within immutable data
structures. If the props of a component are all immutable values,
and the props keep pointing to the same values between renders,
there can be no reason to re-render the component, and it can
be skipped completely.